backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "suchi"
version = "0.1.0"
description = "सूची — CLI-first reference manager with AI-powered research tools"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.9",
    "fastapi>=0.110",
    "uvicorn>=0.27",
    "pyyaml>=6.0",
    "rich>=13.0",
    "httpx>=0.27",
    "pymupdf>=1.24",
    "python-multipart>=0.0.9",
    "tantivy>=0.22",
    "rapidfuzz>=3.0",
    "citeproc-py>=0.6",
]

[project.optional-dependencies]
gdrive = [
    "google-api-python-client>=2.0",
    "google-auth-oauthlib>=1.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.3",
]

[project.scripts]
suchi = "suchi.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/suchi"]

[tool.ruff]
target-version = "py311"
line-length = 100
